requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
httpx==0.28.1
openai==1.59.7
openpyxl==3.1.5
requests==2.32.3
pyyaml==6.0.2
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
google-generativeai==0.8.6
playwright==1.50.0
